Port the original joins repro consolidation
