--- e2e/test/scenarios/joins/reproductions/30743-sort-on-breakout-column-crashes.cy.spec.js
@@ -1,53 +1,0 @@
-import { SAMPLE_DB_ID } from "e2e/support/cypress_data";
-import { SAMPLE_DATABASE } from "e2e/support/cypress_sample_database";
-import {
-  restore,
-  visitQuestionAdhoc,
-  popover,
-  visualize,
-  chartPathWithFillColor,
-} from "e2e/support/helpers";
-
-const { ORDERS, ORDERS_ID, PRODUCTS, PRODUCTS_ID } = SAMPLE_DATABASE;
-
-const query = {
-  dataset_query: {
-    database: SAMPLE_DB_ID,
-    type: "query",
-    query: {
-      "source-table": ORDERS_ID,
-      joins: [
-        {
-          fields: "all",
-          "source-table": PRODUCTS_ID,
-          condition: [
-            "=",
-            ["field", ORDERS.PRODUCT_ID, null],
-            ["field", PRODUCTS.ID, { "join-alias": "Products" }],
-          ],
-          alias: "Products",
-        },
-      ],
-      aggregation: [["count"]],
-      breakout: [["field", PRODUCTS.CATEGORY, { "join-alias": "Products" }]],
-    },
-  },
-};
-
-describe("issue 30743", () => {
-  beforeEach(() => {
-    restore();
-    cy.signInAsAdmin();
-
-    visitQuestionAdhoc(query, { mode: "notebook" });
-  });
-
-  it("should be possible to sort on the breakout column (metabase#30743)", () => {
-    cy.findByLabelText("Sort").click();
-    popover().contains("Category").click();
-
-    visualize();
-    // Check bars count
-    chartPathWithFillColor("#509EE3").should("have.length", 4);
-  });
-});
